refactor(model-proxy): parse-and-call main, testable Proxy entry
`main.swift` held the exit-code taxonomy, a five-branch argument parser,
the logger and the whole `run()` body as top-level code, which a test can
only reach by spawning the binary. Move all of it to a sibling
`Proxy.swift` and leave `main.swift` as the single call into `run()`,
mirroring `TBDHolder`/`Holder.swift` — the precedent that exists for this
exact reason.

No behaviour changes: the text moves verbatim.
